+++ mayan/apps/ocr/tests/test_views.py
@@ -8,18 +8,22 @@ from ..permissions import (
 from ..utils import get_instance_ocr_content
 
 from .literals import TEST_DOCUMENT_VERSION_OCR_CONTENT
-from .mixins import DocumentOCRViewTestMixin, DocumentTypeOCRViewTestMixin
+from .mixins import (
+    DocumentOCRTestMixin, DocumentOCRViewTestMixin,
+    DocumentTypeOCRViewTestMixin
+)
 
 
 class DocumentOCRViewsTestCase(
-    DocumentOCRViewTestMixin, GenericDocumentViewTestCase
+    DocumentOCRTestMixin, DocumentOCRViewTestMixin,
+    GenericDocumentViewTestCase
 ):
     # PyOCR's leak descriptor in get_available_languages and image_to_string
     # Disable descriptor leak test until fixed in upstream
     _skip_file_descriptor_test = True
 
     def test_document_content_delete_view_no_permission(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
 
         response = self._request_test_document_version_ocr_content_delete_view()
         self.assertEqual(response.status_code, 404)
@@ -31,7 +35,8 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_document_content_delete_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document, permission=permission_document_version_ocr
         )
@@ -46,7 +51,8 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_trashed_document_content_delete_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document, permission=permission_document_version_ocr
         )
@@ -63,13 +69,14 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_document_content_view_no_permission(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
 
         response = self._request_test_document_version_ocr_content_view()
         self.assertEqual(response.status_code, 404)
 
     def test_document_content_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document,
             permission=permission_document_version_ocr_content_view
@@ -82,7 +89,8 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_trashed_document_content_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document,
             permission=permission_document_version_ocr_content_view
@@ -94,13 +102,14 @@ class DocumentOCRViewsTestCase(
         self.assertEqual(response.status_code, 404)
 
     def test_document_page_content_view_no_permission(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
 
         response = self._request_test_document_version_page_ocr_content_view()
         self.assertEqual(response.status_code, 404)
 
     def test_document_page_content_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document,
             permission=permission_document_version_ocr_content_view
@@ -113,7 +122,8 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_trashed_document_page_content_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.grant_access(
             obj=self.test_document,
             permission=permission_document_version_ocr_content_view
@@ -200,13 +210,14 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_document_ocr_download_view_no_permission(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
 
         response = self._request_test_document_version_ocr_download_view()
         self.assertEqual(response.status_code, 404)
 
     def test_document_ocr_download_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
+
         self.expected_content_types = ('text/html; charset=utf-8',)
 
         self.grant_access(
@@ -224,7 +235,7 @@ class DocumentOCRViewsTestCase(
         )
 
     def test_trashed_document_ocr_download_view_with_access(self):
-        self.test_document.submit_for_ocr()
+        self._create_test_document_version_ocr_content()
 
         self.grant_access(
             obj=self.test_document,